increase MAX_PORTS to workaround crashes
openoctave ppl report that this fixes the problem for them
In future the MAX_PORTS should be bound to max ports engine parameter of JACK itself.
assert condition like this should not be workarounded but properly checked and error reported.
Kudos for hunting this issue down go to Torben Hohn.