add port id to port name to avoid collisions
This patch changes the pattern how jack port names are derived from alsa
port names when using unique port names. It adds the port id after the
':' to make sure port names are really unique, even when the alsa ports
are not (or too long).
This fixes problems with the ploytec gm5 which has up to 5 in/out ports
and quite long client/port names.